// src/server/generated/voicevox/schema.ts
export interface components {
  schemas: {
// ...
    AudioQuery: {
      /**
       * Accent Phrases
       * @description アクセント句のリスト
       */
      accent_phrases: components["schemas"]["AccentPhrase"][];
      /**
       * Speedscale
       * @description 全体の話速
       */
      speedScale: number;
      /**
       * Pitchscale
       * @description 全体の音高
       */
      pitchScale: number;
      /**
       * Intonationscale
       * @description 全体の抑揚
       */
      intonationScale: number;
      /**
       * Volumescale
       * @description 全体の音量
       */
      volumeScale: number;
      /**
       * Prephonemelength
       * @description 音声の前の無音時間
       */
      prePhonemeLength: number;
      /**
       * Postphonemelength
       * @description 音声の後の無音時間
       */
      postPhonemeLength: number;
      /**
       * Pauselength
       * @description 句読点などの無音時間。nullのときは無視される。デフォルト値はnull
       */
      pauseLength?: number | null;
      /**
       * Pauselengthscale
       * @description 句読点などの無音時間（倍率）。デフォルト値は1
       * @default 1
       */
      pauseLengthScale: number;
      /**
       * Outputsamplingrate
       * @description 音声データの出力サンプリングレート
       */
      outputSamplingRate: number;
      /**
       * Outputstereo
       * @description 音声データをステレオ出力するか否か
       */
      outputStereo: boolean;
      /**
       * Kana
       * @description [読み取り専用]AquesTalk 風記法によるテキスト。音声合成用のクエリとしては無視される
       */
      kana?: string;
    };
// ...
    Speaker: {
      /**
       * Name
       * @description 名前
       */
      name: string;
      /**
       * Speaker Uuid
       * @description キャラクターのUUID
       */
      speaker_uuid: string;
      /**
       * Styles
       * @description スタイルの一覧
       */
      styles: components["schemas"]["SpeakerStyle"][];
      /**
       * Version
       * @description キャラクターのバージョン
       */
      version: string;
      /** @description キャラクターの対応機能 */
      supported_features?: components["schemas"]["SpeakerSupportedFeatures"];
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}